% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-alpinejs.R
\name{convert_from_alpine}
\alias{convert_from_alpine}
\title{Helper to convert JSON coming back from the client}
\usage{
convert_from_alpine(data)
}
\arguments{
\item{data}{result of of an input$<store_name>_data}
}
\description{
Helper to convert JSON coming back from the client
}
