% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-alpinejs.R
\name{update_alpine_data}
\alias{update_alpine_data}
\title{Update data on the client}
\usage{
update_alpine_data(session, name, data)
}
\arguments{
\item{session}{\code{session} from Shiny}

\item{name}{Name of the data store you want to update}

\item{data}{List of data}
}
\description{
Update data on the client
}
