% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-alpinejs.R
\name{withAlpine}
\alias{withAlpine}
\title{Allows the use of R friendly names for Alpine directives
Any HTML attribute staring with x_ is converted to x-
x_on_ is converted to x-on:
x_bind_ is converted to x-bind:
x_shiny_data is converted to x-shiny-data}
\usage{
withAlpine(code)
}
\arguments{
\item{code}{shiny.tag list}
}
\description{
Allows the use of R friendly names for Alpine directives
Any HTML attribute staring with x_ is converted to x-
x_on_ is converted to x-on:
x_bind_ is converted to x-bind:
x_shiny_data is converted to x-shiny-data
}
